remove shallow network module
