+++ datadog_checks_base/datadog_checks/base/utils/http.py
@@ -6,6 +6,7 @@ from __future__ import annotations
 import logging
 import os
 import re
+import socket
 import warnings
 from collections import ChainMap
 from contextlib import ExitStack, contextmanager
@@ -27,7 +28,6 @@ from ..config import is_affirmative
 from ..errors import ConfigurationError
 from .common import ensure_bytes, ensure_unicode
 from .headers import get_default_headers, update_headers
-from .network import create_socket_connection
 from .time import get_timestamp
 from .tls import SUPPORTED_PROTOCOL_VERSIONS, TlsConfig, create_ssl_context
 
@@ -103,6 +103,38 @@ PROXY_SETTINGS_DISABLED = {
 KERBEROS_STRATEGIES = {}
 
 UDS_SCHEME = 'unix'
+
+
+def create_socket_connection(hostname, port=443, sock_type=socket.SOCK_STREAM, timeout=10):
+    """See: https://github.com/python/cpython/blob/40ee9a3640d702bce127e9877c82a99ce817f0d1/Lib/socket.py#L691"""
+    err = None
+    try:
+        for res in socket.getaddrinfo(hostname, port, 0, sock_type):
+            af, socktype, proto, canonname, sa = res
+            sock = None
+            try:
+                sock = socket.socket(af, socktype, proto)
+                sock.settimeout(timeout)
+                sock.connect(sa)
+                # Break explicitly a reference cycle
+                err = None
+                return sock
+
+            except socket.error as _:
+                err = _
+                if sock is not None:
+                    sock.close()
+
+        if err is not None:
+            raise err
+        else:
+            raise socket.error('No valid addresses found, try checking your IPv6 connectivity')  # noqa: G
+    except socket.gaierror as e:
+        err_code, message = e.args
+        if err_code == socket.EAI_NODATA or err_code == socket.EAI_NONAME:
+            raise socket.error('Unable to resolve host, check your DNS: {}'.format(message))  # noqa: G
+
+        raise
 
 
 def get_tls_config_from_options(new_options):
